A little refactoring.

+++ src/order/models.py
@@ -24,7 +24,7 @@ class Order(models.Model):
     phone = PhoneNumberField(verbose_name="Phone Number")
     address = models.CharField(max_length=300, verbose_name="Address")
     city = models.CharField(max_length=300, verbose_name="City")
-    postal_code = models.IntegerField(verbose_name="Postal Code")
+    postal_code = models.BigIntegerField(verbose_name="Postal Code")
     created = models.DateTimeField(auto_now_add=True, verbose_name="Created date and time")
     updated = models.DateTimeField(auto_now=True, verbose_name="Updated date and time")
     paid = models.BooleanField(default=False, verbose_name="Is paid")
